refactor: logs components
