--- src/components/Deploy/DeployLogs/index.tsx
@@ -1,4 +1,0 @@
-export * from './DeploysInfos'
-export * from './DeployLogs'
-export * from './OperationsList'
-export * from './DetailLogView'
