+++ src/components/Deploy/index.ts
@@ -1,4 +1,3 @@
 export { DeployDag } from './DeployDag'
 export { DeployOperations } from './DeployOperations'
-export { DeployLogs } from './DeployLogs'
 export { DeployPreview } from './DeployPreview'
